moves gitlab integration class to its own file cleaning up main file
